Rework models and validation for backward compatibility
