+++ src/ServiceControl.Config/UI/SharedInstanceEditor/ForwardingOption.cs
@@ -1,0 +1,8 @@
+﻿namespace ServiceControl.Config.UI.SharedInstanceEditor
+{
+    public class ForwardingOption
+    {
+        public string Name { get; set; }
+        public bool Value { get; set; }
+    }
+}
